The 64 bit version of the valid move generator isn't better It uses a slimy trick to go off the ends of the area and loses anyway because it ends up doing extra work.  This also gives us a cleaner comparison to the managed code as they are both using 32 bit state machines in the valid move generator now.
The managed code couldn't do the slimy negative offset tricks
that the 64 bit vector required to go both up and down with
twice as many state machines.  If I was doing to do this again
I would try to do more than one 'y' value at a time in parallel.