rename wrath-sharp to wrath-sharp-vec128
The first version I did was the vector128 version which turns out to be really bad
but for historical reasons it had the best name and people often look at it first
which is a horrible idea.

I removed it from the standard benchmark script and renamed the directory to indicate
that it is what it is.